requirements.txt
customtkinter
Pillow
PyYAML
pyinstaller